requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
requests>=2.31.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
sqlalchemy>=2.0.23
alembic>=1.13.1
redis>=5.0.1
celery>=5.3.4
numpy>=1.24.3
pandas>=2.0.3
scikit-learn>=1.3.0
openai>=1.3.0
anthropic>=0.7.0
google-generativeai>=0.3.0
psutil>=5.9.6
psycopg2-binary>=2.9.7
